+++ addon/src/main/java/com/byteowls/vaadin/mediumeditor/MediumEditorState.java
@@ -3,6 +3,8 @@ package com.byteowls.vaadin.mediumeditor;
 import com.byteowls.vaadin.mediumeditor.options.Options;
 import com.vaadin.shared.ui.JavaScriptComponentState;
 
+import elemental.json.JsonValue;
+
 public class MediumEditorState extends JavaScriptComponentState {
 
   private static final long serialVersionUID = -6062251484304417148L;
@@ -11,5 +13,6 @@ public class MediumEditorState extends JavaScriptComponentState {
   public boolean loggingEnabled;
   public boolean focusOutlineEnabled = true;
   public Options options;
+  public JsonValue optionsJson;
 
 }